had to update order and transactional functions

+++ internal/assets/models/holdingModels.go
@@ -8,5 +8,5 @@ type Holding struct {
 	ID        uuid.UUID `db:"id"`
 	AccountID uuid.UUID `db:"account_id"`
 	AssetID   uuid.UUID `db:"asset_id"`
-	Quantity  int       `db:"quantity"`
+	Quantity  float64   `db:"quantity"`
 }
